Benchmark functions

launch functions in background and watch the pid file to get used memory.

This function is best used with long running functions like arrow or duckdb requests that doesn't fit with classic benchmarking methods like utils::Rprof and profmem.

memory is extracted every interval sec in /proc/<pid>/status

start_mem is measured just before launching the function.

max_mem is the max of all measured mem

Usage

timemoir(..., verbose = TRUE, n = 1, interval = 0.1)

Arguments

...

functions to benchmark.

verbose

A boolean. If TRUE (default) print information messages.

n

An integer. number of time each function must run (default 1)

interval

(default 0.1) sleep interval between memory check in sec

Value

A result tibble with one row per benchmarked function and 5 columns:

fname, function name (as string).

duration duration (in sec) of the function or NA if function fails.

error, error message if function fails, NA otherwise.

start_mem memory used before function benchmark (in KB).

max_mem max used memory (in KB).

Examples

timemoir(Sys.sleep(2), Sys.sleep())
#> 
#> 
#> # A tibble: 2 × 5
#>   fname        duration error                                  start_mem max_mem
#>   <chr>           <dbl> <chr>                                      <dbl>   <dbl>
#> 1 Sys.sleep(2)     2.00  NA                                       121336  121948
#> 2 Sys.sleep()     NA    "argument \"time\" is missing, with n…    121336  120128

timemoir(Sys.sleep(1), Sys.sleep(), verbose=FALSE)
#> # A tibble: 2 × 5
#>   fname        duration error                                  start_mem max_mem
#>   <chr>           <dbl> <chr>                                      <dbl>   <dbl>
#> 1 Sys.sleep(1)     1.00  NA                                       121404  122292
#> 2 Sys.sleep()     NA    "argument \"time\" is missing, with n…    121404  120196